Refactor aria utils and added more test cases
