+++ packages/@romejs/compiler/lint/rules/jsx-a11y/noRedundantRoles.test.ts
@@ -14,6 +14,30 @@ test(
 					"<h1 role='heading'></h1>",
 					"<dialog role='dialog'></dialog>",
 					"<input  type='checkbox' role='checkbox' />",
+					"<figure  role='figure'></figure>",
+					"<form  role='form'></form>",
+					"<table  role='grid'></table>",
+					"<td  role='gridcell'></td>",
+					"<fieldset  role='group'></fieldset>",
+					"<img src='foo' alt='bar'  role='img' />",
+					"<a role='link'></a>",
+					"<link role='link' />",
+					"<ol role='list' ></ol>",
+					"<ul role='list' ></ul>",
+					"<select role='listbox' ></select>",
+					"<li role='listitem' ></li>",
+					"<nav role='navigation' ></nav>",
+					"<option role='option' ></option>",
+					"<tr role='row' ></tr>",
+					"<tbody role='rowgroup' ></tbody>",
+					"<tfoot role='rowgroup' ></tfoot>",
+					"<thead role='rowgroup' ></thead>",
+					"<th scope='row' role='rowheader' ></th>",
+					"<input type='search' role='searchbox' />",
+					"<table role='table' ></table>",
+					"<dt role='term' ></dt>",
+					"<textarea role='textbox' ></textarea>",
+					"<input type='text' role='textbox' />",
 				],
 				valid: [
 					"<article role='presentation' ></article>",
