+++ packages/@romejs/compiler/lint/utils/aria/index.ts
@@ -1,0 +1,6 @@
+export {default as isElementInteractive} from "./isElementInteractive";
+export {default as isRoleInteractive} from "./isRoleInteractive";
+export {default as ariaPropsMap} from "./ariaPropsMap";
+export {default as ariaRolesMap} from "./ariaRolesMap";
+export {default as elementsToConceptsMap} from "./elementsToConceptsMap";
+export * from "./types";
