+++ packages/@romejs/compiler/lint/utils/aria/isRoleInteractive.ts
@@ -1,0 +1,5 @@
+import {ARIARoleDefinition} from "@romejs/compiler/lint/utils/aria/types";
+
+export default function isRoleInteractive(role: ARIARoleDefinition) {
+	return role.superClassRole.includes("widget");
+}
